Add new filename field to Buffer type, and filenameE primitive
We can now distinguish between buffers that are connected to a file, and
those that are not (like *sratch* or mini buffers). The consequences are
that buffers have both names (via bufNameE) and filenames (fileNameE).
The name is used internally, the file name is used when writing.

Using fwriteToE on a buffer with no filename sets the filename for that
buffer. It is an error to call fwriteE on a buffer with no filename
connected to it.