Fixed Haskell comment syntax for '------'
Previously if we had a line that consisted only of '-' characters it was caught
as a symbol instead of a comment. For example if you headed a section with:
----------------------
-- Some utilties
Then the top line of this would be coloured in the default style instead of the
comment style as it should be. This is hopefully now fixed.

A further problem was that if you wrote
"--@"
this was as a symbol, but
"--@ " (with the space)
was coloured as a comment.