Fixed modifySelectionB to accurately place the point afterwards
Yi.Region.modifyRegionB had to be altered to return the difference
in length between the old and new regions.

Additionally clarified the movement commands in Yi.Buffer
So there is now a more general moveB which will move left if
the given distance is negative and right otherwise. Of course this means
moveB is just 'rightB' but I think it clarifies it a bit.

Also added a function to return the difference between the point and
the selection point, useful for saying which is the further from the
beginning/end of the buffer.