Fail with proper message when 'hoogle' is not on user's path.
This patch is not ideal in my opinion as the Yi.Process.runProgCommand
does not offer a way to distinguish a command that is not on path from a
command that returned an error code.  However, this is preferable to the
existing state where 'hoogleRaw' returns an empty list when 'hoogle' cmd
fails, causing 'hoogle' function to die because of invalid pattern
match.

Also adding my name to the contributors list.