JS: Beginning work on better stroking.
This is the first step towards correctly error styling stuff which doesn't
look right.  What it does is that it checks  for any error tokens in a list of
them and then depending on whether any of them were errors, returns an
appropriate stroking function for all of the tokens.