Make sure the current buffer and window is the same as that which spawned minibuffer. Resolves issues 293 and 297.
Added the ability for an action to be associated with a buffer such that the action is executed when
the buffer is deleted using Yi.Editor.deleteBuffer. This is used by the minibuffer to assure the
focus is always as expected when the minibuffer is closed. 

I'm only partially satisfied with this solution, but it works and is safer than the other solutions
I was going to apply ;-)