Add vte ui (Gtk with integrated terminal)
Since the Pango UI sorta sucks, I thought this would be a nice way to get a
usable Gtk UI with less work. This will reduce the amount of work that needs
to go into making the UI (since it only needs the vty UI).

I bet there's also a way to use IORef and forking so that we don't have to
launch an entirely separate instance of Yi, but I haven't figured this out
yet.