Don't clear buffer when prompting for file.
Apparently it has been there for multi-item support but there's no
documentation on what such thing is meant to do or why do we need to
show it for ‘""’ considering we call it again soon later. Perhaps this
line was meant to go somewhere else? I'm not sure, I'd love to hear some
reasoning. Perhaps @jyp can explain this behaviour?

For now, this fixes #475 and closes #474.