Prompt the user for a comment string when missing.
Just like in emacs, we let the user pick a string to use for comments if
the mode doesn't already define one. We remember this choice by
modifying the current mode for the buffer on the fly with the comment
syntax the user picked. This change is buffer-local and doesn't affect
any other buffers: we don't want to force the comment string in every
other buffer just because it happened to be using the same mode.

Unfortunately due to suddenly needing the YiM type and a few buffer
functions in scope in Yi.Buffer.Misc, there are quite a lot of boot file
changes but it's either this or having to create a new module in a
rather odd place with no real purpose.

Fixes #413.

Conflicts:
	yi/src/library/Yi/Keymap/Emacs.hs