Expose forkAction, very experimental concurrency
This is probably severely broken but from what I can tell, it works for
the simple case. I could happily edit while something else was going in
the background.

The biggest advantage is that it allows us to run any actions (including
YiA) concurrently, whereas previously even if we forkIO'd, we'd still be
stuck waiting for the result, hanging everything else.

I'm sure there's a good reason for why this was not done before. Perhaps
it's because it uses 'onYiVar' which no doubt reaches out into the sole
MVar.