Make /i have an effect in vim :substitute.
Pretty trivial fix.

Still doesn't support regex, and I'm not sure why we didn't from the
beginning. I'm not up for making a behavior change at the moment,
but I wouldn't be opposed to making regex the default (or at least have
some option to enable it instead of always quoting).