Make sticky eol local to buffer.
Many of our BufferM functions need to be aware of sticky eols
when dealing with block-style regions. Instead of converting
all of those to EditorM, it makes sense to annotate the buffer
with whether or not sticky eol is enabled.

Sticky eol will never be enabled in the emacs map, and the vim
map now uses the BufferM sticky eol instead of the one internal
to its own state.