Support using a different Vty config
I'm not sure if anyone ever used this, but it seems like
a good idea not to regress if possible. I was already looking
at doing this anyway, since the current configuration (before
the package split) was silly and probably why it was removed
in the package split.

By using a YiConfigVariable instance we can implant the value
back in the configuration safely without having a reverse dependency.