Retain range text thing on show for :copy
So that you don't have :<'>'copy, press tab, and get
:copy.

(This was less so a bug in :copy and more so a limitation
of the way parseRange was written to return a BufferM rather
than an AST representing the range. This is elegant but limited
as it can't be turned into a string. We can work around this
in Attoparsec by using match)