Use a real range parser in substitution ex
While :%s/blah/blarg/g is pretty common, so is switching into line
visual mode, and then running a substitute on a small region of a large
file to prevent collateral damage. I was continually bothered by lack
of support for this and finally we have the required bits to fix it
easily.